Refactoring of Database Engines, with huge code consolidation
