+++ app/Modules/Demo/Controllers/Database/Engine.php
@@ -25,19 +25,15 @@ class Engine extends BaseController
         //  Demo 2:
         $demo2_example1 = $engine->selectAll("SELECT * FROM " . DB_PREFIX . "car;");
 
-
-
         // Demo 3
         $data = array('make' => 'BMW', 'model' => 'i8', 'costs' => 138000);
+
         $carid = $engine->insert(DB_PREFIX . 'car', $data);
 
         $demo3_example1 = $carid;
 
-
-
         // Demo 4
         $demo4_example1 = $engine->delete(DB_PREFIX . 'car', array('carid' => $carid));
-
 
         $this->set('demo2_example1', $demo2_example1);
         $this->set('demo3_example1', $demo3_example1);
@@ -55,20 +51,19 @@ class Engine extends BaseController
         // Demo 2:
         $demo2_example1 = $engine->selectAll("SELECT * FROM " . DB_PREFIX . "car;");
 
-
         // Demo 3
         $data = array('make' => 'BMW', 'model' => 'i8', 'costs' => 138000);
+
         $carid = $engine->insert(DB_PREFIX . 'car', $data);
 
         $demo3_example1 = $carid;
 
-
         // Demo 4
         $demo4_example1 = $engine->delete(DB_PREFIX . 'car', array('carid' => $carid));
-
 
         $this->set('demo2_example1', $demo2_example1);
         $this->set('demo3_example1', $demo3_example1);
         $this->set('demo4_example1', $demo4_example1);
     }
+    
 }
